docker/web/app/requirements.txt
Flask
mysql-connector-python
pandas
pymongo